# Refactor in where obtain hash keys by using common function